refactor(tui): extract the pane runtimes into tui::pane
A locally hosted PTY and its agent sampler, a peer-hosted pane and its
input holding, and the layout-control channel with its reservations each
get a file, with the tests that cover them.
