+++ src/tui.rs
@@ -5,6 +5,7 @@ mod debug_log;
 mod geometry;
 mod input;
 mod multi_pane;
+mod pane;
 mod render;
 mod selection;
 mod share;
@@ -20,13 +21,7 @@ use std::{
     fs::OpenOptions,
     io,
     io::Write,
-    path::{Path, PathBuf},
-    sync::{
-        Arc,
-        atomic::{AtomicBool, AtomicU64, Ordering},
-        mpsc::{self as sync_mpsc, Receiver},
-    },
-    thread::{self, JoinHandle},
+    path::PathBuf,
     time::{Duration, Instant},
 };
 use tokio::sync::{mpsc, watch};
@@ -48,10 +43,7 @@ use portable_pty::PtySize;
 use ratatui::{Terminal, TerminalOptions, Viewport, backend::CrosstermBackend, layout::Rect};
 
 use crate::{
-    agent_detect::{
-        AgentKind, AgentScan, AgentState, PaneAgentTracker, ProcessSnapshot, SysinfoSampler,
-        cwd_for_pid, sample_global_snapshot,
-    },
+    agent_detect::{AgentKind, AgentScan, AgentState, cwd_for_pid},
     kitty_keyboard::KittyKeyboardTracker,
     layout::{Axis, LayoutSnapshot, NewPanePosition, PaneId, TabId},
     lease::{IDLE_AFTER, LeaseDecision, LeaseManager, LeaseState},
@@ -65,10 +57,9 @@ use crate::{
     pty_host::PtyHost,
     screen::{GuestScreen, HostScreen, ScreenFrame, SyncGate},
     session::{
-        CoordinatorResponse, GuestEvent, GuestPane, HostControlEvent, HostPaneChannels,
-        LayoutControlEvent, LayoutControlQueueError, PaneLayoutReconciler, PaneServer,
-        SharedLayoutHost, SharedLayoutMember, layout_snapshot_from_state, pane_wire_id,
-        subscribe_pane,
+        CoordinatorResponse, GuestEvent, GuestPane, HostControlEvent, LayoutControlEvent,
+        PaneLayoutReconciler, PaneServer, SharedLayoutHost, SharedLayoutMember,
+        layout_snapshot_from_state, subscribe_pane,
     },
     transport::Transport,
 };
@@ -86,6 +77,15 @@ use input::{
     keys::{PendingEscape, encode_key, encode_paste, is_quit},
 };
 pub use multi_pane::MultiPaneTui;
+pub use pane::local::SharedLocalPane;
+use pane::{
+    control::{PendingCreate, RemoteSubscriptionState, SharedControl},
+    local::{AGENT_SAMPLE_INTERVAL, AGENT_WATCH_INTERVAL, AgentSamplingWorker},
+    remote::{
+        RemoteInput, RemotePaneDrain, SharedRemotePane, lease_allows_held_input,
+        reconcile_remote_control_attempt, remote_input_decision,
+    },
+};
 use render::panes::render_shared_multi_pane;
 pub use render::panes::{render_multi_pane, render_multi_pane_with_copy_feedback};
 use render::vt::{
@@ -116,843 +116,6 @@ pub(crate) const AGENT_OVERLAY_ANIMATION_INTERVAL: Duration = Duration::from_mil
 
 /// The legacy fixed-grid host/guest footer, which has no chords, agents, or share modal.
 const CONTROL_HELP: &str = "Ctrl+ <p> PANE   <t> TAB   <q> QUIT   Option+ <shift> + <↑↓←→> FOCUS";
-
-/// Scan cadence while any pane's agent state is being inferred from output
-/// timing. Inference has to notice silence promptly, so this is the floor.
-const AGENT_SAMPLE_INTERVAL: Duration = Duration::from_secs(1);
-/// Scan cadence when nothing needs inference — no agents at all, or every agent
-/// reporting through a hook. The scan is then only watching for a process to
-/// appear or exit, and a full `sysinfo` refresh of every process on the machine
-/// once a second to do that is most of this process's idle cost.
-const AGENT_WATCH_INTERVAL: Duration = Duration::from_secs(5);
-
-/// One fixed-grid PTY owned by this process. Its watch channels are registered with the pane
-/// service before the layout coordinator is told that the pane is ready.
-pub struct SharedLocalPane {
-    pane_id: PaneId,
-    host: PtyHost,
-    screen: HostScreen,
-    lease: LeaseManager,
-    host_peer_id: Vec<u8>,
-    locked: bool,
-    exited: bool,
-    exit_report_pending: bool,
-    screen_tx: watch::Sender<ScreenFrame>,
-    lease_tx: watch::Sender<LeaseState>,
-    control_tx: mpsc::Sender<HostControlEvent>,
-    control_rx: mpsc::Receiver<HostControlEvent>,
-    agent_tracker: PaneAgentTracker,
-    sync_gate: SyncGate,
-}
-
-impl SharedLocalPane {
-    pub fn spawn(
-        pane_id: PaneId,
-        grid_rows: u16,
-        grid_cols: u16,
-        host_peer_id: Vec<u8>,
-    ) -> Result<Self, Box<dyn Error>> {
-        Self::spawn_with_cwd(pane_id, grid_rows, grid_cols, host_peer_id, None)
-    }
-
-    pub(crate) fn spawn_with_cwd(
-        pane_id: PaneId,
-        grid_rows: u16,
-        grid_cols: u16,
-        host_peer_id: Vec<u8>,
-        cwd: Option<&Path>,
-    ) -> Result<Self, Box<dyn Error>> {
-        let size = PtySize {
-            rows: grid_rows,
-            cols: grid_cols,
-            pixel_width: 0,
-            pixel_height: 0,
-        };
-        let cwd = cwd.filter(|path| path.is_dir());
-        let host = match PtyHost::spawn_default_shell_with_cwd(size, cwd, Some(pane_id)) {
-            Ok(host) => host,
-            Err(_) if cwd.is_some_and(|path| !path.is_dir()) => {
-                PtyHost::spawn_default_shell_with_cwd(size, None, Some(pane_id))?
-            }
-            Err(error) => return Err(error),
-        };
-        let screen = HostScreen::new(grid_rows, grid_cols)?;
-        let (screen_tx, _) = watch::channel(screen.current_frame().clone());
-        let lease = LeaseManager::new(Vec::new(), Instant::now());
-        let (lease_tx, _) = watch::channel(lease.state().clone());
-        let (control_tx, control_rx) = mpsc::channel(256);
-        Ok(Self {
-            pane_id,
-            host,
-            screen,
-            lease,
-            host_peer_id,
-            locked: false,
-            exited: false,
-            exit_report_pending: false,
-            screen_tx,
-            lease_tx,
-            control_tx,
-            control_rx,
-            agent_tracker: PaneAgentTracker::default(),
-            sync_gate: SyncGate::default(),
-        })
-    }
-
-    pub fn channels(&self) -> HostPaneChannels {
-        HostPaneChannels {
-            pane_id: pane_wire_id(self.pane_id),
-            host_peer_id: self.host_peer_id.clone(),
-            screen_rx: self.screen_tx.subscribe(),
-            lease_rx: self.lease_tx.subscribe(),
-            control_tx: self.control_tx.clone(),
-        }
-    }
-
-    fn view_state(&self) -> PaneViewState {
-        PaneViewState {
-            ready: true,
-            controller_peer_id: Some(self.lease.state().controller_peer_id.clone()),
-            controller_active: !self.lease.state().is_idle_at(Instant::now()),
-            scrollback: 0,
-        }
-    }
-
-    fn drain(&mut self) -> Result<LocalPaneDrain, Box<dyn Error>> {
-        let mut changed = false;
-        if !self.exited && self.host.try_wait()? {
-            changed |= self.transition_exited()?;
-        }
-        if let Some(state) = self.lease.clear_if_idle(Instant::now())? {
-            self.lease_tx.send_replace(state);
-            changed = true;
-        }
-        while let Ok(event) = self.control_rx.try_recv() {
-            if self.exited {
-                continue;
-            }
-            match event {
-                HostControlEvent::Input { peer_id, input } => {
-                    if self.locked && peer_id != self.host_peer_id {
-                        continue;
-                    }
-                    if let LeaseDecision::AcceptInput(bytes) =
-                        self.lease
-                            .input(&peer_id, input.lease_epoch, input.data, Instant::now())
-                    {
-                        if self.host.write_input(&bytes).is_err() {
-                            changed |= self.transition_exited()?;
-                            continue;
-                        }
-                        self.lease_tx.send_replace(self.lease.state().clone());
-                        changed = true;
-                    }
-                }
-                HostControlEvent::TakeControl { peer_id, request } => {
-                    if self.locked && peer_id != self.host_peer_id {
-                        continue;
-                    }
-                    let decision = self.lease.take_control(
-                        peer_id,
-                        request.known_lease_epoch,
-                        Instant::now(),
-                    )?;
-                    match decision {
-                        LeaseDecision::Publish(state) => {
-                            self.lease_tx.send_replace(state);
-                            changed = true;
-                        }
-                        // A normal request while the holder is active does not change the lease,
-                        // but the requester needs an authoritative re-publication to clear its
-                        // pending claim and try again after the idle timeout.
-                        LeaseDecision::RejectActiveController => {
-                            self.lease_tx.send_replace(self.lease.state().clone());
-                            changed = true;
-                        }
-                        LeaseDecision::AcceptInput(_)
-                        | LeaseDecision::RejectStaleInput
-                        | LeaseDecision::RejectStaleRequest => {}
-                    }
-                }
-                HostControlEvent::ReleaseControl { peer_id } => {
-                    if self.lease.state().controller_peer_id == peer_id
-                        && let Some(state) = self.lease.clear_controller(Instant::now())?
-                    {
-                        self.lease_tx.send_replace(state);
-                        changed = true;
-                    }
-                }
-            }
-        }
-        let started = Instant::now();
-        let mut pending = Vec::new();
-        for _ in 0..64 {
-            if started.elapsed() >= Duration::from_millis(4) {
-                break;
-            }
-            let Some(bytes) = self.host.try_read_output()? else {
-                break;
-            };
-            pending.extend_from_slice(&bytes);
-        }
-        if !pending.is_empty() {
-            self.agent_tracker
-                .record_output(Instant::now(), unix_ms_now());
-        }
-        let ready = if pending.is_empty() {
-            self.sync_gate.flush_stale(Instant::now())
-        } else {
-            self.sync_gate.feed(&pending, Instant::now())
-        };
-        if !ready.is_empty() {
-            // One parse/snapshot/diff for the whole batch: process_pty clones the
-            // screen per call, so per-chunk calls dominated CPU under output floods.
-            let frame = self.screen.process_pty(&ready)?;
-            // Read after parsing: the bell is a parser callback, so the count only reflects
-            // this batch once the batch has been fed through.
-            if self.screen.take_bell_count() > 0 {
-                self.agent_tracker.record_completion_signal(Instant::now());
-            }
-            if let Some(reply) = self.screen.take_kitty_keyboard_query_reply()
-                && self.host.write_input(&reply).is_err()
-            {
-                changed |= self.transition_exited()?;
-            } else {
-                self.screen_tx.send_replace(frame);
-                changed = true;
-            }
-        }
-        if !self.exited && self.host.output_closed() {
-            changed |= self.transition_exited()?;
-        }
-        Ok(LocalPaneDrain {
-            changed,
-            newly_exited: std::mem::take(&mut self.exit_report_pending),
-        })
-    }
-
-    fn apply_agent_snapshot(&mut self, scan: &AgentScan<'_>, now: Instant) -> bool {
-        let unix_ms_now = unix_ms_now();
-        let before = self.agent_tracker.listed_agent(now, unix_ms_now);
-        let session_child = self.host.process_id();
-        let detected = session_child.and_then(|pid| scan.classify(pid));
-        self.agent_tracker.update(detected, now, unix_ms_now);
-        self.agent_tracker
-            .observe_pane_liveness(session_child.is_some_and(|pid| scan.has_children(pid)), now);
-        self.agent_tracker.listed_agent(now, unix_ms_now) != before
-    }
-
-    /// Whether this pane's reported state comes from output-timing inference
-    /// rather than a producer inside it — the only case that needs the fast
-    /// sampler cadence, since only inference has to notice silence promptly.
-    fn agent_state_is_inferred(&self) -> bool {
-        self.agent_tracker.active_agent.is_some() && !self.agent_tracker.has_owning_push()
-    }
-
-    fn agent_roster_entry(&mut self, now: Instant) -> Option<AgentRosterEntry> {
-        let (agent, state) = self.agent_tracker.listed_agent(now, unix_ms_now())?;
-        Some(AgentRosterEntry {
-            pane_id: self.pane_id,
-            agent_kind: agent.kind.wire_value().into(),
-            cwd: agent.cwd,
-            state: match state {
-                crate::agent_detect::AgentState::Idle => AgentRosterState::Idle as i32,
-                crate::agent_detect::AgentState::Working => AgentRosterState::Working as i32,
-                crate::agent_detect::AgentState::Done => AgentRosterState::Done as i32,
-                crate::agent_detect::AgentState::Pending => AgentRosterState::Pending as i32,
-                crate::agent_detect::AgentState::Error => AgentRosterState::Error as i32,
-            },
-            working_since_unix_ms: if state == crate::agent_detect::AgentState::Working {
-                self.agent_tracker.reported_working_since_unix_ms()
-            } else {
-                0
-            },
-        })
-    }
-
-    fn input(&mut self, bytes: Vec<u8>) -> Result<(), Box<dyn Error>> {
-        if self.exited {
-            return Ok(());
-        }
-        let epoch = self.lease.state().epoch;
-        let decision = self
-            .lease
-            .input(&self.host_peer_id, epoch, bytes, Instant::now());
-        match decision {
-            LeaseDecision::AcceptInput(bytes) => {
-                if self.host.write_input(&bytes).is_err() {
-                    let _ = self.transition_exited()?;
-                }
-                self.lease_tx.send_replace(self.lease.state().clone());
-            }
-            LeaseDecision::Publish(_) => {}
-            LeaseDecision::RejectStaleInput
-            | LeaseDecision::RejectStaleRequest
-            | LeaseDecision::RejectActiveController => {}
-        }
-        Ok(())
-    }
-
-    fn release_controller(&mut self, peer_id: &[u8]) -> Result<bool, Box<dyn Error>> {
-        if self.lease.state().controller_peer_id != peer_id {
-            return Ok(false);
-        }
-        let Some(state) = self.lease.clear_controller(Instant::now())? else {
-            return Ok(false);
-        };
-        self.lease_tx.send_replace(state);
-        Ok(true)
-    }
-
-    fn set_locked(&mut self, locked: bool) -> Result<bool, Box<dyn Error>> {
-        if self.locked == locked {
-            return Ok(false);
-        }
-        self.locked = locked;
-        if locked
-            && self.lease.state().controller_peer_id != self.host_peer_id
-            && let Some(state) = self.lease.clear_controller(Instant::now())?
-        {
-            self.lease_tx.send_replace(state);
-        }
-        Ok(true)
-    }
-
-    fn resize(&mut self, rows: u16, cols: u16) -> Result<(), Box<dyn Error>> {
-        if self.exited {
-            return Ok(());
-        }
-        if self.screen.screen().size() == (rows, cols) {
-            return Ok(());
-        }
-        self.host.resize(PtySize {
-            rows,
-            cols,
-            pixel_width: 0,
-            pixel_height: 0,
-        })?;
-        let frame = self.screen.resize(rows, cols)?;
-        self.screen_tx.send_replace(frame);
-        Ok(())
-    }
-
-    fn shutdown(&mut self) -> Result<(), Box<dyn Error>> {
-        self.host.shutdown()
-    }
-
-    fn transition_exited(&mut self) -> Result<bool, Box<dyn Error>> {
-        if self.exited {
-            return Ok(false);
-        }
-        self.exited = true;
-        self.exit_report_pending = true;
-        while self.control_rx.try_recv().is_ok() {}
-        if let Some(state) = self.lease.clear_controller(Instant::now())? {
-            self.lease_tx.send_replace(state);
-        }
-        Ok(true)
-    }
-
-    fn mark_exited(&mut self) -> Result<(), Box<dyn Error>> {
-        self.transition_exited()?;
-        self.exit_report_pending = false;
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-struct LocalPaneDrain {
-    changed: bool,
-    newly_exited: bool,
-}
-
-struct AgentSamplingWorker {
-    snapshots: Receiver<Vec<ProcessSnapshot>>,
-    stop: Arc<AtomicBool>,
-    interval_ms: Arc<AtomicU64>,
-    join: Option<JoinHandle<()>>,
-}
-
-impl AgentSamplingWorker {
-    fn spawn() -> Self {
-        let (snapshot_tx, snapshots) = sync_mpsc::channel();
-        let stop = Arc::new(AtomicBool::new(false));
-        let interval_ms = Arc::new(AtomicU64::new(AGENT_SAMPLE_INTERVAL.as_millis() as u64));
-        let worker_stop = stop.clone();
-        let worker_interval = interval_ms.clone();
-        let join = thread::spawn(move || {
-            let mut sampler = SysinfoSampler::default();
-            while !worker_stop.load(Ordering::Relaxed) {
-                if snapshot_tx
-                    .send(sample_global_snapshot(&mut sampler))
-                    .is_err()
-                {
-                    break;
-                }
-                thread::sleep(Duration::from_millis(
-                    worker_interval.load(Ordering::Relaxed),
-                ));
-            }
-        });
-        Self {
-            snapshots,
-            stop,
-            interval_ms,
-            join: Some(join),
-        }
-    }
-
-    /// Set how often the global process scan runs.
-    ///
-    /// The scan is the single most expensive thing this process does when
-    /// nothing is happening — a full `sysinfo` refresh of every process on the
-    /// machine, with exe, cmdline and cwd. It only needs the fast cadence when
-    /// a pane's state is being *inferred* from output timing, which is the one
-    /// case that depends on noticing silence promptly. Panes whose agent
-    /// reports through a hook, and panes with no agent at all, only need the
-    /// scan for liveness and for spotting a new launch.
-    fn set_interval(&self, interval: Duration) {
-        self.interval_ms
-            .store(interval.as_millis() as u64, Ordering::Relaxed);
-    }
-
-    fn latest_snapshot(&self) -> Option<Vec<ProcessSnapshot>> {
-        let mut latest = None;
-        while let Ok(snapshot) = self.snapshots.try_recv() {
-            latest = Some(snapshot);
-        }
-        latest
-    }
-
-    fn shutdown(&mut self) {
-        self.stop.store(true, Ordering::Relaxed);
-        if let Some(join) = self.join.take() {
-            let _ = join.join();
-        }
-    }
-}
-
-struct SharedRemotePane {
-    pane: GuestPane,
-    screen: GuestScreen,
-    lease: Option<LeaseState>,
-    last_lease: Instant,
-    pending_control: bool,
-    held_input: Vec<u8>,
-    exited: bool,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum RemotePaneDrain {
-    Unchanged,
-    Changed,
-    Disconnected,
-}
-
-fn lease_allows_held_input(controller_peer_id: &[u8], peer_id: &[u8]) -> bool {
-    controller_peer_id.is_empty() || controller_peer_id == peer_id
-}
-
-fn reconcile_remote_control_attempt(
-    pending_control: &mut bool,
-    held_input: &mut Vec<u8>,
-    controller_peer_id: &[u8],
-    peer_id: &[u8],
-) {
-    *pending_control = false;
-    if !lease_allows_held_input(controller_peer_id, peer_id) {
-        held_input.clear();
-    }
-}
-
-/// What a guest should do with one keystroke aimed at a pane hosted by someone else.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum RemoteInput {
-    /// Send it now, stamped with the lease epoch this guest currently knows.
-    Send,
-    /// Append it to `held_input`: a lease change of ours is already in flight, and
-    /// anything sent before the answer arrives would be stamped with a dead epoch.
-    Hold,
-    /// Ask the host for the pane, then hold this keystroke until it answers.
-    Request,
-    /// Somebody else is actively typing here. Their pane is protected.
-    Ignore,
-}
-
-/// The guest-side input rule, in one place because two loops need it and they had
-/// drifted apart.
-///
-/// The subtle case is `controller_peer_id` being empty -- a free pane, claimed by
-/// typing into it. That claim is not free: the host bumps the lease epoch the moment it
-/// accepts the first byte, so every keystroke typed during the round trip that follows
-/// arrives stamped with an epoch the host has already left behind and is rejected as
-/// stale. On loopback the window is under a millisecond and no scenario ever caught it;
-/// across an 85ms internet path it silently swallows the next five characters, so
-/// `echo N-CROSS` reaches the shell as `e-CROSS`. Holding the rest of the burst until
-/// the new lease lands costs nothing and keeps the line intact.
-fn remote_input_decision(
-    controller_peer_id: &[u8],
-    peer_id: &[u8],
-    pending_control: bool,
-    held_input_empty: bool,
-    controller_idle: bool,
-) -> RemoteInput {
-    if controller_peer_id == peer_id {
-        return if held_input_empty {
-            RemoteInput::Send
-        } else {
-            RemoteInput::Hold
-        };
-    }
-    if controller_peer_id.is_empty() {
-        return if pending_control {
-            RemoteInput::Hold
-        } else {
-            RemoteInput::Send
-        };
-    }
-    if controller_idle {
-        return if pending_control {
-            RemoteInput::Hold
-        } else {
-            RemoteInput::Request
-        };
-    }
-    RemoteInput::Ignore
-}
-
-impl SharedRemotePane {
-    fn new(pane: GuestPane) -> Self {
-        Self {
-            pane,
-            screen: GuestScreen::new(),
-            lease: None,
-            last_lease: Instant::now(),
-            pending_control: false,
-            held_input: Vec::new(),
-            exited: false,
-        }
-    }
-
-    fn view_state(&self) -> PaneViewState {
-        PaneViewState {
-            ready: self.screen.screen().is_some() && self.lease.is_some(),
-            controller_peer_id: self
-                .lease
-                .as_ref()
-                .map(|lease| lease.controller_peer_id.clone()),
-            controller_active: self
-                .lease
-                .as_ref()
-                .is_some_and(|lease| !lease.is_idle_at(Instant::now())),
-            scrollback: 0,
-        }
-    }
-
-    fn drain(&mut self) -> RemotePaneDrain {
-        if self.exited {
-            while self.pane.events.try_recv().is_ok() {}
-            return RemotePaneDrain::Unchanged;
-        }
-        let mut changed = false;
-        let mut received_lease = false;
-        loop {
-            match self.pane.events.try_recv() {
-                Ok(GuestEvent::ScreenSnapshot(snapshot)) => {
-                    if self
-                        .screen
-                        .apply_snapshot(snapshot.sequence, &snapshot.screen)
-                        .is_ok()
-                    {
-                        self.screen
-                            .set_kitty_keyboard_active(snapshot.kitty_keyboard_active);
-                        changed = true;
-                    }
-                }
-                Ok(GuestEvent::ScreenDelta(delta)) => {
-                    if self
-                        .screen
-                        .apply_delta(delta.base_sequence, delta.sequence, &delta.changes)
-                        .is_ok()
-                    {
-                        self.screen
-                            .set_kitty_keyboard_active(delta.kitty_keyboard_active);
-                        changed = true;
-                    }
-                }
-                Ok(GuestEvent::Lease(lease)) => {
-                    received_lease = true;
-                    self.lease = Some(LeaseState {
-                        controller_peer_id: lease.controller_peer_id,
-                        epoch: lease.lease_epoch,
-                        last_activity: Instant::now(),
-                    });
-                    self.last_lease = Instant::now();
-                    self.pending_control = false;
-                    changed = true;
-                }
-                Ok(GuestEvent::ScreenGap { .. }) => {}
-                Ok(GuestEvent::Disconnected)
-                | Err(tokio::sync::mpsc::error::TryRecvError::Disconnected) => {
-                    return RemotePaneDrain::Disconnected;
-                }
-                Err(tokio::sync::mpsc::error::TryRecvError::Empty) => break,
-            }
-        }
-        if received_lease && let Some(lease) = self.lease.as_ref() {
-            reconcile_remote_control_attempt(
-                &mut self.pending_control,
-                &mut self.held_input,
-                &lease.controller_peer_id,
-                self.pane.controls.peer_id(),
-            );
-        }
-        if !self.pending_control
-            && !self.held_input.is_empty()
-            && self.lease.as_ref().is_some_and(|lease| {
-                lease_allows_held_input(&lease.controller_peer_id, self.pane.controls.peer_id())
-            })
-        {
-            let bytes = std::mem::take(&mut self.held_input);
-            if self
-                .pane
-                .controls
-                .try_input(
-                    self.lease.as_ref().expect("checked lease").epoch,
-                    bytes.clone(),
-                )
-                .is_err()
-            {
-                self.held_input = bytes;
-            }
-        }
-        if changed {
-            RemotePaneDrain::Changed
-        } else {
-            RemotePaneDrain::Unchanged
-        }
-    }
-
-    fn input(&mut self, bytes: Vec<u8>) {
-        if self.exited {
-            return;
-        }
-        let Some(lease) = self.lease.as_ref() else {
-            return;
-        };
-        let claiming_free_pane = lease.controller_peer_id.is_empty();
-        match remote_input_decision(
-            &lease.controller_peer_id,
-            self.pane.controls.peer_id(),
-            self.pending_control,
-            self.held_input.is_empty(),
-            self.last_lease.elapsed() >= IDLE_AFTER,
-        ) {
-            RemoteInput::Send => {
-                if self.pane.controls.try_input(lease.epoch, bytes).is_ok() && claiming_free_pane {
-                    // The host will answer this byte with a new epoch; hold the rest of
-                    // the burst until it does.
-                    self.pending_control = true;
-                }
-            }
-            RemoteInput::Hold => self.held_input.extend_from_slice(&bytes),
-            RemoteInput::Request => {
-                self.held_input.extend_from_slice(&bytes);
-                self.pending_control = self.pane.controls.try_take_control(lease.epoch).is_ok();
-                if !self.pending_control {
-                    self.held_input.clear();
-                }
-            }
-            RemoteInput::Ignore => {}
-        }
-    }
-
-    fn release_controller(&mut self) -> bool {
-        let Some(lease) = self.lease.as_mut() else {
-            return false;
-        };
-        if lease.controller_peer_id != self.pane.controls.peer_id()
-            || self.pane.controls.try_release_control().is_err()
-        {
-            return false;
-        }
-        lease.controller_peer_id.clear();
-        self.pending_control = false;
-        self.held_input.clear();
-        true
-    }
-
-    fn mark_exited(&mut self) {
-        self.exited = true;
-        self.pending_control = false;
-        self.held_input.clear();
-    }
-}
-
-enum SharedControl {
-    Host(SharedLayoutHost),
-    Member(SharedLayoutMember),
-}
-
-impl SharedControl {
-    fn peer_id(&self) -> Vec<u8> {
-        match self {
-            Self::Host(host) => host.ticket().endpoint_addr().id.as_bytes().to_vec(),
-            Self::Member(member) => member.peer_id.clone(),
-        }
-    }
-
-    fn try_request(&self, request: LayoutRequest) -> Result<Option<CoordinatorResponse>, String> {
-        match self {
-            Self::Host(host) => host
-                .handle_local_request(request)
-                .map(Some)
-                .map_err(|error| error.to_string()),
-            Self::Member(member) => member
-                .try_request(request)
-                .map(|()| None)
-                .map_err(layout_queue_message),
-        }
-    }
-
-    fn try_ready(&self, ready: PaneReady) -> Result<Option<CoordinatorResponse>, String> {
-        match self {
-            Self::Host(host) => host
-                .handle_local_ready(ready)
-                .map(Some)
-                .map_err(|error| error.to_string()),
-            Self::Member(member) => member
-                .try_ready(ready)
-                .map(|()| None)
-                .map_err(layout_queue_message),
-        }
-    }
-
-    fn try_failed(&self, failed: PaneFailed) -> Result<(), String> {
-        match self {
-            Self::Host(host) => host
-                .handle_local_failed(failed)
-                .map(|_| ())
-                .map_err(|error| error.to_string()),
-            Self::Member(member) => member.try_failed(failed).map_err(layout_queue_message),
-        }
-    }
-
-    fn try_agent_roster(&self, roster: AgentRoster) -> Result<(), String> {
-        match self {
-            Self::Host(host) => host
-                .publish_local_agent_roster(roster)
-                .map_err(|error| error.to_string()),
-            Self::Member(member) => member
-                .try_agent_roster(roster)
-                .map_err(layout_queue_message),
-        }
-    }
-
-    fn try_event(&mut self, current_revision: u64) -> Option<LayoutControlEvent> {
-        match self {
-            // The coordinator is the authority, so it does not receive its own broadcasts. Poll
-            // its tiny in-memory snapshot to observe joins, departures, and member-originated
-            // commits without adding a second internal control stream.
-            Self::Host(host) => host
-                .session_snapshot()
-                .ok()
-                .filter(|snapshot| {
-                    snapshot
-                        .state
-                        .as_ref()
-                        .is_some_and(|state| state.revision > current_revision)
-                })
-                .map(LayoutControlEvent::Snapshot),
-            Self::Member(member) => member.events.try_recv().ok(),
-        }
-    }
-
-    async fn shutdown(self) {
-        match self {
-            Self::Host(host) => host.close().await,
-            Self::Member(member) => member.shutdown().await,
-        }
-    }
-}
-
-fn layout_queue_message(error: LayoutControlQueueError) -> String {
-    match error {
-        LayoutControlQueueError::Full => String::from("layout request queue is busy"),
-        LayoutControlQueueError::Closed => String::from("layout coordinator disconnected"),
-    }
-}
-
-struct PendingCreate {
-    request_id: u64,
-    base_revision: u64,
-    grid_rows: u16,
-    grid_cols: u16,
-    cwd: Option<PathBuf>,
-}
-
-/// Pure retry state for direct remote-pane subscriptions. Ticks are supplied by the UI drain
-/// loop, which keeps retry behavior deterministic in tests and prevents a failed connect from
-/// either becoming permanent or spinning a busy loop.
-#[derive(Default)]
-struct RemoteSubscriptionState {
-    in_flight: BTreeSet<PaneId>,
-    retry_at: BTreeMap<PaneId, u64>,
-    failures: BTreeMap<PaneId, u8>,
-}
-
-impl RemoteSubscriptionState {
-    const MAX_BACKOFF_TICKS: u64 = 32;
-
-    fn start(&mut self, pane_id: PaneId, tick: u64) -> bool {
-        if self.in_flight.contains(&pane_id)
-            || self
-                .retry_at
-                .get(&pane_id)
-                .is_some_and(|retry_at| *retry_at > tick)
-        {
-            return false;
-        }
-        self.in_flight.insert(pane_id);
-        true
-    }
-
-    fn succeeded(&mut self, pane_id: PaneId) {
-        self.in_flight.remove(&pane_id);
-        self.retry_at.remove(&pane_id);
-        self.failures.remove(&pane_id);
-    }
-
-    fn failed(&mut self, pane_id: PaneId, tick: u64) {
-        self.in_flight.remove(&pane_id);
-        let failures = self.failures.entry(pane_id).or_default();
-        *failures = failures.saturating_add(1);
-        let delay = 1_u64
-            .checked_shl(u32::from((*failures).saturating_sub(1)))
-            .unwrap_or(Self::MAX_BACKOFF_TICKS)
-            .min(Self::MAX_BACKOFF_TICKS);
-        self.retry_at.insert(pane_id, tick.saturating_add(delay));
-    }
-
-    fn nudge(&mut self) {
-        for retry_at in self.retry_at.values_mut() {
-            *retry_at = 0;
-        }
-    }
-
-    fn retain(&mut self, pane_ids: &BTreeSet<PaneId>) {
-        self.in_flight.retain(|pane_id| pane_ids.contains(pane_id));
-        self.retry_at
-            .retain(|pane_id, _| pane_ids.contains(pane_id));
-        self.failures
-            .retain(|pane_id, _| pane_ids.contains(pane_id));
-    }
-
-    #[cfg(test)]
-    fn next_retry_at(&self, pane_id: PaneId) -> Option<u64> {
-        self.retry_at.get(&pane_id).copied()
-    }
-}
 
 /// Blocking terminal runtime for the shared layout. Network tasks keep streams independent while
 /// this loop only drains ready channels and renders the current fixed grids.
@@ -3160,44 +2323,22 @@ mod tests {
 
     use crate::agent_detect::cwd_for_pid;
     use crate::layout::{Axis, NewPanePosition};
-    use crate::lease::{IDLE_AFTER, LeaseManager, LeaseState};
+    use crate::lease::{LeaseManager, LeaseState};
     use crate::screen::HostScreen;
     use crate::{
         protocol::PaneDescriptor,
-        session::{HostSession, SharedLayoutHost, layout_snapshot_from_state},
+        session::{
+            HostPaneChannels, HostSession, SharedLayoutHost, layout_snapshot_from_state,
+            pane_wire_id,
+        },
         transport::{ALPN, Transport},
     };
     use iroh::{Endpoint, RelayMode, endpoint::presets};
 
     use super::{
-        HostControlEvent, HostPaneChannels, HostPaneRuntime, LayoutControlEvent, RemoteInput,
-        RemoteSubscriptionState, SharedLayoutRuntime, SharedLocalPane, UiIntent,
-        lease_allows_held_input, member_label, pane_wire_id, reconcile_remote_control_attempt,
-        remote_input_decision,
+        HostPaneRuntime, LayoutControlEvent, SharedLayoutRuntime, SharedLocalPane, UiIntent,
+        member_label,
     };
-
-    #[test]
-    fn failed_remote_subscription_retries_with_bounded_backoff_and_commit_nudge() {
-        let mut subscriptions = RemoteSubscriptionState::default();
-        assert!(subscriptions.start(7, 0));
-        subscriptions.failed(7, 0);
-        assert!(!subscriptions.start(7, 0));
-        assert!(subscriptions.start(7, 1));
-        subscriptions.failed(7, 1);
-        assert!(!subscriptions.start(7, 2));
-        assert!(subscriptions.start(7, 3));
-
-        for tick in [3, 7, 15, 31, 63, 95] {
-            subscriptions.failed(7, tick);
-            assert!(
-                subscriptions
-                    .next_retry_at(7)
-                    .is_some_and(|next| next <= tick + 32)
-            );
-        }
-        subscriptions.nudge();
-        assert!(subscriptions.start(7, 95));
-    }
 
     async fn loopback_transport() -> Transport {
         let endpoint = Endpoint::builder(presets::Minimal)
@@ -3562,236 +2703,6 @@ mod tests {
     }
 
     #[test]
-    fn rejected_busy_claim_republishes_lease_and_idle_claim_succeeds() {
-        let owner = b"owner".to_vec();
-        let requester = b"guest".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, owner.clone()).expect("local pane");
-        let mut lease_rx = pane.lease_tx.subscribe();
-        pane.lease = LeaseManager::new(owner.clone(), Instant::now());
-        pane.control_tx
-            .try_send(HostControlEvent::TakeControl {
-                peer_id: requester.clone(),
-                request: crate::protocol::TakeControl {
-                    pane_id: pane_wire_id(99),
-                    requester_peer_id: requester.clone(),
-                    known_lease_epoch: 1,
-                },
-            })
-            .expect("busy takeover event");
-        pane.drain().expect("drain busy takeover");
-        assert!(lease_rx.has_changed().expect("lease watch"));
-        assert_eq!(lease_rx.borrow_and_update().controller_peer_id, owner);
-
-        pane.lease = LeaseManager::with_epoch_for_test(
-            pane.host_peer_id.clone(),
-            1,
-            Instant::now() - IDLE_AFTER,
-        );
-        pane.control_tx
-            .try_send(HostControlEvent::TakeControl {
-                peer_id: requester.clone(),
-                request: crate::protocol::TakeControl {
-                    pane_id: pane_wire_id(99),
-                    requester_peer_id: requester.clone(),
-                    known_lease_epoch: 2,
-                },
-            })
-            .expect("idle takeover event");
-        pane.drain().expect("drain idle takeover");
-        assert_eq!(pane.lease.state().controller_peer_id, requester);
-        assert_eq!(pane.lease.state().epoch, 3);
-    }
-
-    #[test]
-    fn idle_controller_is_cleared_and_published_to_watchers() {
-        let host_id = b"host".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, host_id.clone()).expect("local pane");
-        let mut lease_rx = pane.lease_tx.subscribe();
-        pane.lease = LeaseManager::with_epoch_for_test(host_id, 1, Instant::now() - IDLE_AFTER);
-
-        pane.drain().expect("drain idle controller");
-
-        assert!(lease_rx.has_changed().expect("lease watch"));
-        let lease = lease_rx.borrow_and_update().clone();
-        assert!(lease.controller_peer_id.is_empty());
-        assert_eq!(lease.epoch, 2);
-    }
-
-    #[test]
-    fn exited_local_pane_clears_lease_and_rejects_all_controls() {
-        let host_id = b"host".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, host_id.clone()).expect("local pane");
-        let lease_rx = pane.lease_tx.subscribe();
-        pane.lease = LeaseManager::new(host_id.clone(), Instant::now());
-        assert!(pane.transition_exited().expect("exit transition"));
-        assert!(pane.exited);
-        assert!(pane.lease.state().controller_peer_id.is_empty());
-        assert!(lease_rx.has_changed().expect("lease publication"));
-
-        pane.control_tx
-            .try_send(HostControlEvent::TakeControl {
-                peer_id: host_id,
-                request: crate::protocol::TakeControl {
-                    pane_id: pane_wire_id(99),
-                    requester_peer_id: b"host".to_vec(),
-                    known_lease_epoch: pane.lease.state().epoch,
-                },
-            })
-            .expect("queued control");
-        let drained = pane.drain().expect("drain exited pane");
-        assert!(drained.newly_exited);
-        assert!(pane.lease.state().controller_peer_id.is_empty());
-        pane.shutdown().expect("shutdown exited pane");
-    }
-
-    #[test]
-    fn locked_local_pane_rejects_guest_control_and_accepts_host_control() {
-        let host_id = b"host".to_vec();
-        let guest_id = b"guest".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, host_id.clone()).expect("local pane");
-        pane.set_locked(true).expect("lock pane");
-
-        pane.control_tx
-            .try_send(HostControlEvent::TakeControl {
-                peer_id: guest_id.clone(),
-                request: crate::protocol::TakeControl {
-                    pane_id: pane_wire_id(99),
-                    requester_peer_id: guest_id.clone(),
-                    known_lease_epoch: 1,
-                },
-            })
-            .expect("guest takeover event");
-        pane.drain().expect("drain guest takeover");
-        assert!(pane.lease.state().controller_peer_id.is_empty());
-
-        pane.control_tx
-            .try_send(HostControlEvent::Input {
-                peer_id: guest_id.clone(),
-                input: crate::protocol::Input {
-                    pane_id: pane_wire_id(99),
-                    lease_epoch: 1,
-                    data: b"blocked".to_vec(),
-                },
-            })
-            .expect("guest input event");
-        pane.drain().expect("drain guest input");
-        assert!(pane.lease.state().controller_peer_id.is_empty());
-
-        pane.control_tx
-            .try_send(HostControlEvent::TakeControl {
-                peer_id: host_id.clone(),
-                request: crate::protocol::TakeControl {
-                    pane_id: pane_wire_id(99),
-                    requester_peer_id: host_id.clone(),
-                    known_lease_epoch: 1,
-                },
-            })
-            .expect("host takeover event");
-        pane.drain().expect("drain host takeover");
-        assert_eq!(pane.lease.state().controller_peer_id, host_id);
-        pane.shutdown().expect("shutdown local pane");
-    }
-
-    #[test]
-    fn locking_local_pane_clears_guest_lease_but_keeps_host_lease() {
-        let host_id = b"host".to_vec();
-        let guest_id = b"guest".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, host_id.clone()).expect("local pane");
-        let mut lease_rx = pane.lease_tx.subscribe();
-        pane.lease = LeaseManager::new(guest_id, Instant::now());
-
-        pane.set_locked(true).expect("lock pane");
-        assert!(lease_rx.has_changed().expect("lease watch"));
-        assert!(lease_rx.borrow_and_update().controller_peer_id.is_empty());
-
-        pane.lease = LeaseManager::new(host_id.clone(), Instant::now());
-        pane.set_locked(false).expect("unlock pane");
-        pane.set_locked(true).expect("relock pane");
-        assert_eq!(pane.lease.state().controller_peer_id, host_id);
-        pane.shutdown().expect("shutdown local pane");
-    }
-
-    #[test]
-    fn remote_release_clears_the_host_lease_immediately() {
-        let host_id = b"host".to_vec();
-        let guest_id = b"guest".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, host_id).expect("local pane");
-        let mut lease_rx = pane.lease_tx.subscribe();
-        pane.lease = LeaseManager::new(guest_id.clone(), Instant::now());
-        pane.control_tx
-            .try_send(HostControlEvent::ReleaseControl { peer_id: guest_id })
-            .expect("remote release event");
-
-        pane.drain().expect("drain remote release");
-
-        assert!(lease_rx.has_changed().expect("lease watch"));
-        let lease = lease_rx.borrow_and_update().clone();
-        assert!(lease.controller_peer_id.is_empty());
-        assert_eq!(lease.epoch, 2);
-
-        pane.shutdown().expect("shutdown local pane");
-    }
-
-    #[test]
-    fn local_input_crossing_the_idle_timeout_reclaims_and_delivers_it() {
-        let host_id = b"host".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, host_id.clone()).expect("local pane");
-        let mut lease_rx = pane.lease_tx.subscribe();
-        pane.lease =
-            LeaseManager::with_epoch_for_test(b"remote".to_vec(), 1, Instant::now() - IDLE_AFTER);
-
-        pane.input(b"printf boundary-input-delivered\\n".to_vec())
-            .expect("first input at idle boundary");
-
-        assert_eq!(pane.lease.state().controller_peer_id, host_id);
-        assert_eq!(pane.lease.state().epoch, 3);
-        assert!(lease_rx.has_changed().expect("lease watch"));
-        let lease = lease_rx.borrow_and_update().clone();
-        assert_eq!(lease.controller_peer_id, b"host");
-        assert_eq!(lease.epoch, 3);
-
-        let deadline = Instant::now() + Duration::from_secs(2);
-        let mut output = Vec::new();
-        while Instant::now() < deadline {
-            while let Some(bytes) = pane.host.try_read_output().expect("PTY reader") {
-                output.extend(bytes);
-            }
-            if String::from_utf8_lossy(&output).contains("boundary-input-delivered") {
-                pane.shutdown().expect("shutdown local pane");
-                return;
-            }
-            thread::sleep(Duration::from_millis(10));
-        }
-        pane.shutdown().expect("shutdown local pane");
-        panic!(
-            "first input was not delivered to the PTY: {:?}",
-            String::from_utf8_lossy(&output)
-        );
-    }
-
-    #[test]
-    fn new_local_pane_starts_free_while_the_host_retains_pty_ownership() {
-        let host_id = b"host".to_vec();
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, host_id.clone()).expect("local pane");
-
-        assert!(pane.lease.state().controller_peer_id.is_empty());
-        assert_eq!(pane.host_peer_id, host_id);
-
-        pane.shutdown().expect("shutdown local pane");
-    }
-
-    #[test]
-    fn local_pane_resize_publishes_a_replacement_screen_frame() {
-        let mut pane = SharedLocalPane::spawn(99, 1, 1, b"host".to_vec()).expect("pane");
-        let before = pane.screen.current_frame().sequence;
-        pane.resize(3, 4).expect("resize");
-        assert_eq!(pane.screen.screen().size(), (3, 4));
-        assert!(pane.screen.current_frame().sequence > before);
-        assert_eq!(pane.screen.current_frame().base_sequence, 0);
-        pane.shutdown().expect("shutdown");
-    }
-
-    #[test]
     fn new_host_runtime_starts_free_while_the_host_retains_pty_ownership() {
         let host_id = b"host".to_vec();
         let screen = HostScreen::new(1, 1).expect("screen");
@@ -3822,73 +2733,6 @@ mod tests {
         assert!(lease_rx.borrow().controller_peer_id.is_empty());
 
         runtime.host.shutdown().expect("shutdown host runtime");
-    }
-
-    #[test]
-    fn remote_held_input_follows_the_final_lease_owner() {
-        let peer = b"requester";
-        let mut pending_control = true;
-        let mut held_input = b"first".to_vec();
-
-        assert!(lease_allows_held_input(b"", peer));
-        reconcile_remote_control_attempt(&mut pending_control, &mut held_input, b"", peer);
-        assert!(!pending_control, "free leases release the retry gate");
-        assert_eq!(held_input, b"first", "free leases retain queued input");
-
-        reconcile_remote_control_attempt(&mut pending_control, &mut held_input, b"other", peer);
-        assert!(
-            held_input.is_empty(),
-            "a later controller wins and discards stale input"
-        );
-    }
-
-    #[test]
-    fn claiming_a_free_remote_pane_holds_the_rest_of_the_burst() {
-        let peer = b"typist";
-
-        // The first keystroke goes out and claims the pane.
-        assert_eq!(
-            remote_input_decision(b"", peer, false, true, false),
-            RemoteInput::Send
-        );
-        // Everything typed during the round trip has to wait: the host has already
-        // bumped the epoch, so sending now means the host drops those bytes as stale.
-        // This is the whole bug -- invisible on loopback, five lost characters at 85ms.
-        assert_eq!(
-            remote_input_decision(b"", peer, true, true, false),
-            RemoteInput::Hold
-        );
-    }
-
-    #[test]
-    fn remote_input_respects_the_controller_lease() {
-        let peer = b"typist";
-
-        assert_eq!(
-            remote_input_decision(peer, peer, false, true, false),
-            RemoteInput::Send,
-            "the controller types straight through"
-        );
-        assert_eq!(
-            remote_input_decision(peer, peer, false, false, false),
-            RemoteInput::Hold,
-            "queued input keeps its place in line"
-        );
-        assert_eq!(
-            remote_input_decision(b"other", peer, false, true, false),
-            RemoteInput::Ignore,
-            "an active controller is protected from takeover"
-        );
-        assert_eq!(
-            remote_input_decision(b"other", peer, false, true, true),
-            RemoteInput::Request,
-            "an idle controller can be displaced"
-        );
-        assert_eq!(
-            remote_input_decision(b"other", peer, true, true, true),
-            RemoteInput::Hold,
-            "one request at a time; the rest of the burst waits for the answer"
-        );
     }
 
     #[test]
